+++ src/main/java/chess/domain/ChessGame.java
@@ -4,6 +4,7 @@ import chess.domain.command.Command;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceFactory;
 import chess.domain.piece.Pieces;
+import chess.domain.piece.info.Color;
 import chess.domain.position.Position;
 
 public class ChessGame {
