+++ src/main/java/chess/domain/piece/Pieces.java
@@ -1,7 +1,7 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
-import chess.domain.Score;
+import chess.domain.piece.info.Color;
+import chess.domain.piece.info.Score;
 import chess.domain.position.Position;
 
 import java.util.ArrayList;
