+++ src/main/java/chess/domain/piece/info/Color.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece.info;
 
 public enum Color {
     WHITE,
