+++ src/main/java/chess/view/OutputView.java
@@ -1,12 +1,11 @@
 package chess.view;
 
-import chess.domain.Color;
 import chess.domain.ScoreStatus;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Pieces;
+import chess.domain.piece.info.Color;
 
-import static chess.domain.position.Position.CHANGE_LINE_POINT;
-import static chess.domain.position.Position.POSITIONS;
+import static chess.domain.position.Position.*;
 
 public class OutputView {
     private OutputView() {
