+++ src/test/java/chess/domain/ScoreStatusTest.java
@@ -2,6 +2,7 @@ package chess.domain;
 
 import chess.domain.piece.PieceFactory;
 import chess.domain.piece.Pieces;
+import chess.domain.piece.info.Color;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
