+++ src/test/java/chess/domain/ScoreTest.java
@@ -1,5 +1,6 @@
 package chess.domain;
 
+import chess.domain.piece.info.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
